refactor out Dir from Zipper
